Remove proxy graphs
After some reflection, it was discovered that proxy graphs are just
context graphs with an arbitrary restriction: if an edge exists in one
context, it must exist in all contexts that contain the endpoints (via
other edges).

Additionally, the previously removed concept of "overtrees" was a
further refinement of the concept of proxy graphs: the additional
restriction being that a node can only exist in at most one context.